Make AgentLauncher extend HealthIndicator
Force all implementations of the Agent Launcher interface to report health metrics for their underlying implementation.

Remove redundant LocalAgentLauncherHealthIndicator and roll logic into LocalAgentLauncherImpl
